removed newlines, moved Usergroups object to correct package
