+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/group/usergroups/UsergroupsCreateExtension.kt
@@ -1,7 +1,6 @@
 package io.olaph.slack.dto.jackson.group.usergroups
 
 import io.olaph.slack.dto.jackson.common.types.Channel
-import io.olaph.slack.dto.jackson.common.types.Usergroup
 import io.olaph.slack.dto.jackson.group.channels.sample
 
 fun SuccessfulUsergroupsCreateResponse.Companion.sample() = SuccessfulUsergroupsCreateResponse(true, Usergroup.sample())
