+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/group/usergroups/UsergroupsExtension.kt
@@ -1,9 +1,5 @@
 package io.olaph.slack.dto.jackson.group.usergroups
 
-import io.olaph.slack.dto.jackson.common.types.AutoType
-import io.olaph.slack.dto.jackson.common.types.Prefs
-import io.olaph.slack.dto.jackson.common.types.Usergroup
-
 fun Usergroup.Companion.sample() = Usergroup(id = "",
         teamId = "",
         isUsergroup = true,
