+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/group/usergroups/UsergroupsListExtension.kt
@@ -1,7 +1,5 @@
 package io.olaph.slack.dto.jackson.group.usergroups
 
-import io.olaph.slack.dto.jackson.common.types.Usergroup
-
 fun SuccessfulUsergroupsListResponse.Companion.sample() = SuccessfulUsergroupsListResponse(true, listOf(Usergroup.sample()))
 
 fun ErrorUsergroupsListResponse.Companion.sample() = ErrorUsergroupsListResponse(false, "")
