+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/io/olaph/slack/dto/jackson/group/usergroups/users/UsergroupsUsersUpdateExtension.kt
@@ -1,6 +1,6 @@
 package io.olaph.slack.dto.jackson.group.usergroups.users
 
-import io.olaph.slack.dto.jackson.common.types.Usergroup
+import io.olaph.slack.dto.jackson.group.usergroups.Usergroup
 import io.olaph.slack.dto.jackson.group.usergroups.sample
 
 fun SuccessfulUsergroupUsersUpdateResponse.Companion.sample() = SuccessfulUsergroupUsersUpdateResponse(true, Usergroup.sample())
